Fixes missing function
